Module Res

Global module for resizable datastructures and default implementations

Default strategies

module DefStrat : sig ... end

Default strategy for resizable datastructures

module BitDefStrat : sig ... end

Same as DefStrat, but the minimum size is 1024 elements (bits).

Default instantiation of standard resizable datastructures

module Array : sig ... end

Resizable parameterized array using the default reallocation strategy.

module Floats : sig ... end

Resizable float array using the default reallocation strategy.

module Bits : sig ... end

Resizable bit vector using the default reallocation strategy.

module Weak : sig ... end

Resizable weak array using the default reallocation strategy.

module Buffer : sig ... end

Resizable buffer using the default reallocation strategy.

Functors for creating standard resizable datastructures from strategies

module MakeArray (S : sig ... end) : sig ... end

Functor that creates resizable parameterized arrays from reallocation strategies.

module MakeFloats (S : sig ... end) : sig ... end

Functor that creates resizable float arrays from reallocation strategies.

module MakeBits (S : sig ... end) : sig ... end

Functor that creates resizable bit vectors from reallocation strategies.

module MakeWeak (S : sig ... end) : sig ... end

Functor that creates resizable weak arrays from reallocation strategies.

module MakeBuffer (S : sig ... end) : sig ... end

Functor that creates resizable buffers (=string arrays) from reallocation strategies.